1222 종목정보 입력중

+++ src/main/java/kr/co/kfs/assetedu/repository/Com01CorpRepository.java
@@ -7,7 +7,7 @@ import kr.co.kfs.assetedu.model.QueryAttr;
 
 public interface Com01CorpRepository {
 	List<Com01Corp> selectList (QueryAttr queryAttr);
-	Integer selectCount(QueryAttr queryAttr);
+	Long selectCount(QueryAttr queryAttr);
 	Com01Corp selectOne(Com01Corp corp);
